# pkg/query: Use maps instead of slices for children for O(1) access